% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfrs.R
\name{hfrs}
\alias{hfrs}
\title{HFRS function
This function calculates the Hospital Frailty Risk Score from ICD codes}
\usage{
hfrs(df, columns)
}
\arguments{
\item{df}{The data frame containing ICD codes}

\item{columns}{array identifying the columns containing ICD codes}
}
\description{
HFRS function
This function calculates the Hospital Frailty Risk Score from ICD codes
}
\examples{
data$HFRS<-hfrs(dataframe, c("D1","D2","D3","D4","D5"))
hfrs(dataframe,1:20)
hfrs()
}
\keyword{frailty,}
\keyword{hfrs}
\keyword{risk,}
